moved locationData to LocationCards improving performance for displaying LocationCards on Homepage
